refactor: reorganize scratch org files
