+++ src/org/scratch/errorCodes.ts
@@ -6,12 +6,12 @@
  */
 
 import { Optional } from '@salesforce/ts-types';
-import { Messages } from '../messages';
-import { SfError } from '../sfError';
-import { Logger } from '../logger/logger';
-import { ScratchOrgInfo } from './scratchOrgTypes';
+import { Messages } from '../../messages';
+import { SfError } from '../../sfError';
+import { Logger } from '../../logger/logger';
+import { ScratchOrgInfo } from './types';
 import { ScratchOrgCache } from './scratchOrgCache';
-import { emit } from './scratchOrgLifecycleEvents';
+import { emit } from './lifecycle';
 
 const WORKSPACE_CONFIG_FILENAME = 'sfdx-project.json';
