+++ src/org/scratch/lifecycle.ts
@@ -4,9 +4,9 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { Lifecycle } from '../lifecycleEvents';
-import { AuthFields } from './authInfo';
-import { ScratchOrgInfo } from './scratchOrgTypes';
+import { Lifecycle } from '../../lifecycleEvents';
+import { AuthFields } from '../authInfo';
+import { ScratchOrgInfo } from './types';
 
 export const scratchOrgLifecycleEventName = 'scratchOrgLifecycleEvent';
 export const scratchOrgLifecycleStages = [
