+++ src/org/scratch/settingsGenerator.ts
@@ -9,17 +9,17 @@ import * as path from 'node:path';
 import { isEmpty, env, upperFirst, Duration } from '@salesforce/kit';
 import { ensureObject, JsonMap } from '@salesforce/ts-types';
 import * as js2xmlparser from 'js2xmlparser';
-import { Logger } from '../logger/logger';
-import { SfError } from '../sfError';
-import { StructuredWriter } from '../util/structuredWriter';
-import { StatusResult } from '../status/types';
-import { PollingClient } from '../status/pollingClient';
-import { ZipWriter } from '../util/zipWriter';
-import { DirectoryWriter } from '../util/directoryWriter';
-import { Lifecycle } from '../lifecycleEvents';
-import { Messages } from '../messages';
-import { ScratchOrgInfo, ObjectSetting } from './scratchOrgTypes';
-import { Org } from './org';
+import { Logger } from '../../logger/logger';
+import { SfError } from '../../sfError';
+import { StructuredWriter } from '../../util/structuredWriter';
+import { StatusResult } from '../../status/types';
+import { PollingClient } from '../../status/pollingClient';
+import { ZipWriter } from '../../util/zipWriter';
+import { DirectoryWriter } from '../../util/directoryWriter';
+import { Lifecycle } from '../../lifecycleEvents';
+import { Messages } from '../../messages';
+import { Org } from '../org';
+import { ScratchOrgInfo, ObjectSetting } from './types';
 
 Messages.importMessagesDirectory(__dirname);
