+++ test/unit/org/scratchOrgInfoGeneratorTest.ts
@@ -17,7 +17,7 @@ import {
   getAncestorIds,
   generateScratchOrgInfo,
   getScratchOrgInfoPayload,
-} from '../../../src/org/scratchOrgInfoGenerator';
+} from '../../../src/org/scratch/scratchOrgInfoGenerator';
 import { SfProjectJson } from '../../../src/sfProject';
 import { Messages } from '../../../src/messages';
